Move HaxballEvents file to the modles folder
