+++ src-ts/index.ts
@@ -9,9 +9,9 @@ export default function haxilium<TPlayer extends Player>(config: RoomConfig<TPla
 
 export { Event } from './decorators/Event';
 export { Module } from './decorators/Module';
-export { HaxballEventsInterface as HaxballEvents } from './HaxballEvents';
 export { Scores } from './interfaces/Scores';
 export { Team } from './interfaces/Team';
 export { Vector } from './interfaces/Vector';
+export { HaxballEventsInterface as HaxballEvents } from './models/HaxballEvents';
 export { Player } from './models/Player';
 export { Room } from './Room';
