+++ src-ts/interfaces/Module.ts
@@ -1,4 +1,4 @@
-import { HaxballEvents } from '../HaxballEvents';
+import { HaxballEvents } from '../models/HaxballEvents';
 import { Player } from '../models/Player';
 
 export interface Module<TPlayer extends Player> extends Partial<HaxballEvents<TPlayer>> {
